requirements.txt
openai
camoufox[geoip]
mcp[cli]
python-dotenv
python-telegram-bot  # optional: for --telegram mode
